test: scope flow handoff integration to postgres
