+++ internal/api/integration_test/project_test.go
@@ -3,10 +3,8 @@
 package integration_test
 
 import (
-	"strings"
 	"testing"
 
-	"github.com/go-faster/jx"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	api "github.com/zitadel/nextgen/api/generated"
@@ -90,48 +88,6 @@ func TestCreateProjectProvisionsDefaultLoginFlow(t *testing.T) {
 	assert.Equal(t, "identifier", flow.Response.Step.Name)
 	assert.Contains(t, flow.Response.Step.Fields, "email")
 	assert.Contains(t, flow.Response.Step.Actions, "submit")
-
-	afterIdentifierResp, err := client.SubmitFlowStep(t.Context(), &api.FlowSubmitRequest{
-		Action: "submit",
-		Fields: api.NewOptFlowSubmitRequestFields(api.FlowSubmitRequestFields{
-			"email": jx.Raw(`"fresh@example.test"`),
-		}),
-	}, api.SubmitFlowStepParams{
-		ID:    flow.Response.ID,
-		Zflow: zflowCookieValue(t, flow.SetCookie.Value),
-	})
-	require.NoError(t, err)
-	afterIdentifier, ok := afterIdentifierResp.(*api.SubmitFlowStepOK)
-	require.True(t, ok, helpers.MustMarshal(t, afterIdentifierResp))
-	assert.Equal(t, "register", afterIdentifier.Response.Step.Name)
-	assert.Contains(t, afterIdentifier.Response.Step.Fields, "password")
-
-	doneResp, err := client.SubmitFlowStep(t.Context(), &api.FlowSubmitRequest{
-		Action: "submit",
-		Fields: api.NewOptFlowSubmitRequestFields(api.FlowSubmitRequestFields{
-			"email":    jx.Raw(`"fresh@example.test"`),
-			"password": jx.Raw(`"Correct-Horse-42!"`),
-		}),
-	}, api.SubmitFlowStepParams{
-		ID:    afterIdentifier.Response.ID,
-		Zflow: zflowCookieValue(t, afterIdentifier.SetCookie.Value),
-	})
-	require.NoError(t, err)
-	done, ok := doneResp.(*api.SubmitFlowStepOK)
-	require.True(t, ok, helpers.MustMarshal(t, doneResp))
-	assert.Equal(t, "done", done.Response.Step.Name)
-	assert.NotEmpty(t, done.Response.HandoffToken.Value)
-	assert.True(t, done.Response.HandoffTokenExpiresAt.Set)
-}
-
-func zflowCookieValue(t *testing.T, setCookie string) string {
-	t.Helper()
-	pair := strings.Split(setCookie, ";")[0]
-	name, value, ok := strings.Cut(pair, "=")
-	require.True(t, ok, "set-cookie did not contain a cookie pair: %q", setCookie)
-	require.Equal(t, "_zflow", name)
-	require.NotEmpty(t, value)
-	return value
 }
 
 func TestGetProject(t *testing.T) {
